% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{ls.dimensionality}
\alias{ls.dimensionality}
\title{List the dimensionality measures}
\usage{
ls.dimensionality()
}
\value{
A list of dimensionality measures names
}
\description{
List the dimensionality measures
}
\examples{
ls.dimensionality()
}
