% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{ls.linearity}
\alias{ls.linearity}
\title{List the linearity measures}
\usage{
ls.linearity()
}
\value{
A list of linearity measures names
}
\description{
List the linearity measures
}
\examples{
ls.linearity()
}
