% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{ls.neighborhood}
\alias{ls.neighborhood}
\title{List the neighborhood measures}
\usage{
ls.neighborhood()
}
\value{
A list of neighborhood measures names
}
\description{
List the neighborhood measures
}
\examples{
ls.neighborhood()
}
