% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{ls.overlapping}
\alias{ls.overlapping}
\title{List the overlapping measures}
\usage{
ls.overlapping()
}
\value{
A list of overlapping measures names
}
\description{
List the overlapping measures
}
\examples{
ls.overlapping()
}
