% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{smoothness}
\alias{smoothness}
\alias{smoothness.default}
\alias{smoothness.formula}
\title{Smoothness measures}
\usage{
smoothness(...)

\method{smoothness}{default}(x, y, measures = "all", ...)

\method{smoothness}{formula}(formula, data, measures = "all", ...)
}
\arguments{
\item{...}{Not used.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{measures}{A list of measures names or \code{"all"} to include all them.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class.}
}
\value{
A list named by the requested regression smoothness measure.
}
\description{
In regression problems, the smoother the function to be fitted to the data, 
the simpler it shall be. Larger variations in the inputs and/or outputs, on 
the other hand, usually indicate the existence of more intricate 
relationships between them.
}
\details{
The following measures are allowed for this method:
 \describe{
   \item{"S1"}{Output distribution (S1)}
   \item{"S2"}{Input distribution (S2)}
   \item{"S3"}{Error of a nearest neighbor regressor (S3)}
   \item{"S4"}{Non-linearity of nearest neighbor regressor (S4)}
 }
}
\examples{
## Extract all smoothness measures
data(cars)
smoothness(speed~., cars)
}
\references{
Ana C Lorena and Aron I Maciel and Pericles B C Miranda and Ivan G Costa and
   Ricardo B C Prudencio. (2018). Data complexity meta-features for 
   regression problems. Machine Learning, 107, 1, 209--246.
}
\seealso{
Other complexity-measures: \code{\link{balance}},
  \code{\link{correlation}}, \code{\link{dimensionality}},
  \code{\link{linearity.class}},
  \code{\link{linearity.regr}}, \code{\link{neighborhood}},
  \code{\link{network}}, \code{\link{overlapping}}
}
